Create useTableColumnsFromLocalStorage util fn, shared defaultColumns config.
