+++ src/uniref/config/UniRefColumnConfiguration.tsx
@@ -20,6 +20,16 @@ export enum UniRefColumn {
   created = 'created',
 }
 
+export const defaultColumns = [
+  UniRefColumn.id,
+  UniRefColumn.name,
+  UniRefColumn.types,
+  UniRefColumn.count,
+  UniRefColumn.organism,
+  UniRefColumn.length,
+  UniRefColumn.identity,
+];
+
 export const UniRefColumnConfiguration = new Map<
   UniRefColumn,
   {
